Move talk adapter to new file.
