+++ source/src/in/joind/EventTalks.java
@@ -334,294 +334,3 @@ public class EventTalks extends JIActivity implements OnClickListener {
     }
 }
 
-
-/**
- * Adapter that hold our talk rows. See  JIEventAdapter class in main.java for more info
- */
-class JITalkAdapter extends ArrayAdapter<JSONObject> implements Filterable {
-    private ArrayList<JSONObject> items, filtered_items;
-    private Context context;
-    private TimeZone tz;
-    private StarredFilter filter;
-    private boolean isAuthenticated;
-
-    public JITalkAdapter(Context context, int textViewResourceId, ArrayList<JSONObject> mTalks, TimeZone tz, boolean isAuthenticated) {
-        super(context, textViewResourceId, mTalks);
-        this.context = context;
-        this.items = mTalks;
-        this.filtered_items = mTalks;
-        this.tz = tz;
-        this.isAuthenticated = isAuthenticated;
-    }
-
-    public View getView(int position, View convertview, ViewGroup parent) {
-        View v = convertview;
-        if (v == null) {
-            LayoutInflater vi = (LayoutInflater) this.context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-            v = vi.inflate(R.layout.talkrow, null);
-        }
-
-        final JSONObject o = filtered_items.get(position);
-        if (o == null) return v;
-
-        // Convert the supplied date/time string into something we can use
-        Date talkDate = null;
-        SimpleDateFormat outputTalkDateFormat = null;
-        try {
-            SimpleDateFormat inputTalkDateFormat = new SimpleDateFormat(context.getString(R.string.apiDateFormat));
-            talkDate = inputTalkDateFormat.parse(o.getString("start_date"));
-            String fmt = Build.VERSION.SDK_INT <= 8 ? "E d MMM yyyy" : "E d LLL yyyy";
-            outputTalkDateFormat = new SimpleDateFormat(fmt + ", HH:mm");
-            outputTalkDateFormat.setTimeZone(tz);
-        } catch (Exception e) {
-            e.printStackTrace();
-            // Nothing here. Date is probably formatted badly
-        }
-        long cts = System.currentTimeMillis() / 1000;
-
-        // Set a bit of darker color when the talk is currently held (the date_given is less than an hour old)
-        if (talkDate != null && cts - talkDate.getTime() <= 3600 && cts - talkDate.getTime() >= 0) {
-            v.setBackgroundColor(Color.rgb(218, 218, 204));
-        } else {
-            // This isn't right. We shouldn't set a white color, but the default color
-            v.setBackgroundColor(Color.rgb(255, 255, 255));
-        }
-
-        String t2Text;
-        int commentCount = o.optInt("comment_count");
-        if (commentCount == 1) {
-            t2Text = String.format(this.context.getString(R.string.generalCommentSingular), commentCount);
-        } else {
-            t2Text = String.format(this.context.getString(R.string.generalCommentPlural), commentCount);
-        }
-
-        String track = "";
-        try {
-            track = o.optJSONArray("tracks").getJSONObject(0).optString("track_name");
-        } catch (JSONException e) {
-            // Ignore if no track is available
-        }
-
-        TextView t1 = (TextView) v.findViewById(R.id.TalkRowCaption);
-        TextView t2 = (TextView) v.findViewById(R.id.TalkRowComments);
-        TextView t3 = (TextView) v.findViewById(R.id.TalkRowSpeaker);
-        TextView t4 = (TextView) v.findViewById(R.id.TalkRowTime);
-        TextView t5 = (TextView) v.findViewById(R.id.TalkRowTrack);
-        if (t1 != null) t1.setText(o.optString("talk_title"));
-        if (t2 != null) t2.setText(t2Text);
-        if (t4 != null) t4.setText(outputTalkDateFormat.format(talkDate));
-        if (t5 != null) t5.setText(track);
-
-        // Speaker details
-        ArrayList<String> speakerNames = new ArrayList<String>();
-        try {
-            JSONArray speakerEntries = o.getJSONArray("speakers");
-            for (int i = 0; i < speakerEntries.length(); i++) {
-                speakerNames.add(speakerEntries.getJSONObject(i).getString("speaker_name"));
-            }
-        } catch (JSONException e) {
-            Log.d(JIActivity.LOG_JOINDIN_APP, "Couldn't get speaker names");
-            e.printStackTrace();
-        }
-        if (speakerNames.size() == 1) {
-            t3.setText("Speaker: " + speakerNames.get(0));
-        }
-        else if (speakerNames.size() > 1) {
-            String allSpeakers = TextUtils.join(", ", speakerNames);
-            t3.setText("Speakers: " + allSpeakers);
-        }
-        else {
-            t3.setText("");
-        }
-
-        // Set specified talk category image
-        Resources resources = context.getResources();
-        if (o.optString("type").compareTo("Talk") == 0) t1.setCompoundDrawables(resources.getDrawable(R.drawable.talk), null, null, null);
-        if (o.optString("type").compareTo("Social Event") == 0) t1.setCompoundDrawables(resources.getDrawable(R.drawable.socialevent), null, null, null);
-        if (o.optString("type").compareTo("Workshop") == 0) t1.setCompoundDrawables(resources.getDrawable(R.drawable.workshop), null, null, null);
-        if (o.optString("type").compareTo("Keynote") == 0) t1.setCompoundDrawables(resources.getDrawable(R.drawable.keynote), null, null, null);
-
-        ImageView rateview = (ImageView) v.findViewById(R.id.TalkRowRating);
-        int rate = o.optInt("average_rating", 0);
-        switch (rate) {
-            case 0:
-                rateview.setBackgroundResource(R.drawable.rating_0);
-                break;
-            case 1:
-                rateview.setBackgroundResource(R.drawable.rating_1);
-                break;
-            case 2:
-                rateview.setBackgroundResource(R.drawable.rating_2);
-                break;
-            case 3:
-                rateview.setBackgroundResource(R.drawable.rating_3);
-                break;
-            case 4:
-                rateview.setBackgroundResource(R.drawable.rating_4);
-                break;
-            case 5:
-                rateview.setBackgroundResource(R.drawable.rating_5);
-                break;
-        }
-
-        // Show/hide the starred icon if the talk is starred
-        boolean starredStatus = o.optBoolean("starred", false);
-        CheckBox starredImageButton = (CheckBox) v.findViewById(R.id.TalkRowStarred);
-        starredImageButton.setVisibility(isAuthenticated ? View.VISIBLE : View.GONE);
-        starredImageButton.setChecked(starredStatus);
-        final View finalV = v;
-        starredImageButton.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                markTalkStarred(finalV, o.optString("starred_uri"), ((CheckBox) view).isChecked());
-            }
-        });
-
-        return v;
-    }
-
-    /**
-     * Mark the talk as starred - update the icon and submit the request
-     * @param isStarred
-     */
-    protected void markTalkStarred(final View parentRow, final String starredURI, final boolean isStarred) {
-        final CheckBox starredImageButton = (CheckBox) parentRow.findViewById(R.id.TalkRowStarred);
-        final ProgressBar progressBar = (ProgressBar) parentRow.findViewById(R.id.TalkRowProgress);
-
-        new Thread() {
-            public void run() {
-                updateProgressStatus(progressBar, starredImageButton, true);
-
-                final String result = doStarTalk(isStarred, starredURI);
-
-                updateProgressStatus(progressBar, starredImageButton, false);
-            }
-        }.start();
-    }
-
-    /**
-     * CALLED FROM SEPARATE THREAD
-     * This shows/hides the progressbar and the checkbox alternately.
-     *
-     * @param progressBar
-     * @param starredImageButton
-     * @param showProgress
-     */
-    private void updateProgressStatus(final ProgressBar progressBar, final CheckBox starredImageButton, final boolean showProgress) {
-        progressBar.post(new Runnable() {
-            @Override
-            public void run() {
-                progressBar.setVisibility(showProgress ? View.VISIBLE : View.GONE);
-            }
-        });
-        starredImageButton.post(new Runnable() {
-            @Override
-            public void run() {
-                starredImageButton.setVisibility(showProgress ? View.GONE : View.VISIBLE);
-            }
-        });
-    }
-
-    /**
-     * Post/delete the starred status from this talk
-     *
-     * @param initialState
-     * @return
-     */
-    private String doStarTalk(boolean initialState, String starredURI) {
-        JIRest rest = new JIRest(context);
-        int error = rest.requestToFullURI(starredURI, null, initialState ? JIRest.METHOD_POST : JIRest.METHOD_DELETE);
-
-        if (error != JIRest.OK) {
-            return String.format(context.getString(R.string.generalStarringError), rest.getError());
-        }
-
-        // Everything went as expected
-        if (initialState) {
-            return context.getString(R.string.generalSuccessStarred);
-        } else {
-            return context.getString(R.string.generalSuccessUnstarred);
-        }
-    }
-
-    public Filter getFilter() {
-        if (filter == null) {
-            filter = new StarredFilter();
-        }
-        return filter;
-    }
-
-    public int getCount() {
-        return filtered_items.size();
-    }
-
-    public JSONObject getItem(int position) {
-        return filtered_items.get(position);
-    }
-
-    /**
-     * Starred filter
-     */
-    public class StarredFilter extends Filter {
-        private boolean checkStarredStatus = false;
-
-        public void setCheckStarredStatus(boolean checkStarredStatus) {
-            this.checkStarredStatus = checkStarredStatus;
-        }
-
-        @SuppressWarnings("unchecked")
-        protected void publishResults(CharSequence prefix, FilterResults results) {
-            filtered_items = (ArrayList<JSONObject>) results.values;
-
-            notifyDataSetChanged();
-        }
-
-        protected FilterResults performFiltering(CharSequence match) {
-            FilterResults results = new FilterResults();
-            ArrayList<JSONObject> i = new ArrayList<JSONObject>();
-
-            // Multiple filters here
-            if (checkStarredStatus || (match != null && match.length() > 0)) {
-                for (int index = 0; index < items.size(); index++) {
-                    JSONObject json = items.get(index);
-                    JSONArray tracks = json.optJSONArray("tracks");
-                    if (tracks.length() == 0) {
-                        continue;
-                    }
-
-                    int tracksLength = tracks.length();
-                    for (int j = 0; j < tracksLength; j++) {
-                        JSONObject track = tracks.optJSONObject(j);
-                        if (track == null) {
-                            continue;
-                        }
-
-                        // Add to the filtered result list when the match is present in the URI
-                        // If we need to check starred status as well, let's do that too
-                        if (match.toString().length() > 0) {
-                            // We have a track to match against, this happens first
-                            if (track.optString("track_uri").equals(match.toString())) {
-                                if ((checkStarredStatus && json.optBoolean("starred")) || !checkStarredStatus) {
-                                    i.add(json);
-                                    break;
-                                }
-                            }
-                        } else if (checkStarredStatus && json.optBoolean("starred")) {
-                            i.add(json);
-                            break;
-                        }
-                    }
-                }
-                results.values = i;
-                results.count = i.size();
-            } else {
-                synchronized (items) {
-                    results.values = items;
-                    results.count = items.size();
-                }
-            }
-
-            return results;
-        }
-    }
-}
